+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Refactor `UiNode::trace` to work with widget and list nodes too.
+
 * Refactor `actual_size` and related `actual_*` properties to get the size on layout.
     - Before the size was sampled on render, now it is sampled before.
     - This change means any state hooked to the actual size will now update before the new size is visible on the next frame.