+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+
+# 0.16.5
+
 * Fix race condition in `zng::task::SignalOnce`.
 
 * Event/command notify requested during a layout pass now all run before the next render pass.